added fade in/out to intro. fixed a few warnings

+++ src/GameObject/MainMenuBackgroundRenderer.h
@@ -53,8 +53,8 @@ public:
         }
 
         if(sweepAnimation < numberColumns * 4){
-            int xSweep = sweepAnimation / 2;
-            int ySweep = 0;
+            size_t xSweep = sweepAnimation / 2;
+            size_t ySweep = 0;
             while(xSweep < numbers[0].size() * 2){
                 if(ySweep < numbers.size() && xSweep < numbers[0].size()){
                     numbers[ySweep][xSweep] = RandomGenerator::getInstance().getRandom() > 0.5 ? 1 : 0;
